Add files via upload
Exploratory analysis of Nvidia data on jupyter notebooks. I downloaded yfinance, NVDA data which included the Date, Open, High, Low, Close, Adj Close, Volume.
I loaded the data that was already in a data frame, checked for missing values, checked for the information of the data types present. I used the function describe() to find the mean, standard deviation, percentiles, max and min values of the data to better understand the structure of the data. I also created a line graph of the Close column to better see the time series structure of the data. I plotted the 30-day moving average and compared it to the line graph of the Close column.