Add files via upload
After finding the information using the describe function, the standard deviation value was huge in all the columns, and this led me to analyze the distribution of the data. A high standard deviation both on the Close price and the Volume column shows a great degree of variability and volatility of the data overtime.
First using the mean and the standard deviation of the Close price column, I created o normal distribution with the same length as that of my close price data. The purpose of this was to compare the normally distributed data with the original nvidia_data. I found that my data was skewed to the right with relatively heavy tails.
A further Shapiro-Wilk test showed that the nvidia_data is not normally distributed as the p_value being 0.00 is less than alpha. The distribution is most likely to be a leptokurtic distribution.